requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
transformers>=4.41.0
datasets>=2.20.0
accelerate>=0.31.0
evaluate>=0.4.2
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
rich>=13.9.0
sentencepiece
kaggle>=1.6.14
joblib
pyyaml